add utils mod to eth_vault
